+++ kfft_rader.h
@@ -3,4 +3,4 @@
 #include "kfft_guts.h"
 
 void
-kf_rader(kfft_cpx* Fout, const size_t fstride, const kfft_kplan_t* st, int m, int p);
+kf_rader(kfft_cpx* Fout, const uint32_t fstride, const kfft_kplan_t* st, uint32_t m, uint32_t p);
